Minor fixes
- Wrong assert for failed transfer
- Datatype for unsigned type in loop
- Duplicated declaration in tests

+++ src/commands.cpp
@@ -1208,7 +1208,7 @@ bool CommandPutNodes::procresult(Result r)
             vector<NewNode> emptyVec;
             client->app->putnodes_result(r.errorOrOK(), type, emptyVec);
 
-            for (int i=0; i < nn.size(); i++)
+            for (size_t i = 0; i < nn.size(); i++)
             {
                 nn[i].localnode.reset();
             }
